+++ packages/vike/node/runtime/html/injectAssets/getHtmlTags.ts
@@ -23,7 +23,7 @@ import { getConfigDefinedAt } from '../../../../shared/page-configs/getConfigDef
 import { htmlElementId_globalContext, htmlElementId_pageContext } from '../../../../shared/htmlElementIds.js'
 import { isFontFallback } from '../../renderPage/isFontFallback.js'
 import type { PageContextServer } from '../../../../types/PageContext.js'
-import { inferNonceAttr, type PageContextCspNonce } from '../../renderPage/csp.js'
+import { inferNonceAttr, type PageContextCspNonce } from '../../csp.js'
 
 const stamp = '__injectFilterEntry'
