+++ packages/vike/node/runtime/html/injectAssets/inferHtmlTags.ts
@@ -5,7 +5,7 @@ export { scriptAttrs }
 
 import { assert } from '../../utils.js'
 import type { PageAsset } from '../../renderPage/getPageAssets.js'
-import { inferNonceAttr, PageContextCspNonce } from '../../renderPage/csp.js'
+import { inferNonceAttr, PageContextCspNonce } from '../../csp.js'
 
 // TODO/now rename scriptAttrs scriptCommonAttrs
 // We can't use `defer` here. With `defer`, the entry script won't start before `</body>` has been parsed, preventing progressive hydration during SSR streaming, see https://github.com/vikejs/vike/pull/1271
